yolo5/requirements.txt
pyyaml
loguru
boto3
requests
